support auto generate static for gaussian (gaussian_random) (#52422)
* support auto generate static for gaussian (gaussian_random)

* move gaussian_random_batch_size_like Kernels from gaussian_random_op.* to gaussian_random_batch_size_like_op.*
